betting_api/requirements.txt
django
djangorestframework
pygments
MySQL-python
Pillow